--- src/main/java/domain/Height.java
@@ -1,13 +1,0 @@
-package domain;
-
-public class Height {
-    private final int height;
-
-    public Height(int heightInput) {
-        this.height = heightInput;
-    }
-
-    public int getHeight() {
-        return height;
-    }
-}
