+++ src/main/java/domain/user/Users.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.user;
 
 import java.util.List;
 import java.util.stream.Collectors;
